refactor: Move sync method back to api factory
