+++ lib/contentful.js
@@ -9,7 +9,6 @@ import version from '../version'
 import createHttpClient from 'contentful-sdk-core/create-http-client'
 import createContentfulApi from './create-contentful-api'
 import entities from './entities'
-import createSyncRequester from './create-sync-requester'
 import createLinkResolver from './create-link-resolver'
 
 /**
@@ -51,13 +50,9 @@ export default function createClient (axios, params) {
   })
 
   const http = createHttpClient(axios, params)
-  const api = createContentfulApi({
+  return createContentfulApi({
     http: http,
     entities: entities,
     shouldLinksResolve: shouldLinksResolve
   })
-
-  api.sync = createSyncRequester(http, shouldLinksResolve)
-
-  return api
 }
