+++ lib/create-contentful-api.js
@@ -8,6 +8,7 @@
  */
 
 import createRequestConfig from 'contentful-sdk-core/create-request-config'
+import pagedSync from './paged-sync'
 
 /**
  * Creates API object with methods to access functionality common to Contentful's
@@ -149,6 +150,26 @@ export default function createContentfulApi ({
     .then((response) => wrapAssetCollection(response.data), errorHandler)
   }
 
+  /**
+   * Synchronizes either all the content or only new content since last sync
+   * See <a href="https://www.contentful.com/developers/docs/concepts/sync/">Synchronization</a> for more information.
+   * @memberof ContentfulAPIClient
+   * @param  {Object} query - Query object for the sync call. One of initial or nextSyncToken always needs to be specified, but not both.
+   * @param  {boolean?} query.initial - Indicates if this is the first sync. Use it if you don't have a sync token.
+   * @param  {string?} query.nextSyncToken - The token you got the last time you used this method. Ensures you only get changed content.
+   * @param  {string=} query.type - Filter by this type (Entry or Asset)
+   * @param  {string=} query.content_type - Filter by this content type id
+   * @param  {boolean=} query.resolveLinks - When true, links to other Entries or Assets are resolved. Default: true.
+   * @return {Promise<Sync.SyncCollection>} Promise for the collection resulting of a sync operation
+   * @example
+   * client.sync()
+   * .then((response) => console.log(response.entries, response.assets, response.nextSyncToken))
+   */
+  function sync (query = {}) {
+    const resolveLinks = shouldLinksResolve(query)
+    return pagedSync(http, query, resolveLinks)
+  }
+
   return {
     getSpace: getSpace,
     getContentType: getContentType,
@@ -156,7 +177,8 @@ export default function createContentfulApi ({
     getEntry: getEntry,
     getEntries: getEntries,
     getAsset: getAsset,
-    getAssets: getAssets
+    getAssets: getAssets,
+    sync: sync
   }
 }
