+++ lib/paged-sync.js
@@ -35,28 +35,6 @@ import mixinToPlainObject from 'contentful-sdk-core/mixins/to-plain-object'
  * @type Entities.Asset
  */
 
-export default function createSyncRequester (http, shouldLinksResolve) {
-  /**
-   * Synchronizes either all the content or only new content since last sync
-   * See <a href="https://www.contentful.com/developers/docs/concepts/sync/">Synchronization</a> for more information.
-   * @memberof ContentfulAPIClient
-   * @param  {Object} query - Query object for the sync call. One of initial or nextSyncToken always needs to be specified, but not both.
-   * @param  {boolean?} query.initial - Indicates if this is the first sync. Use it if you don't have a sync token.
-   * @param  {string?} query.nextSyncToken - The token you got the last time you used this method. Ensures you only get changed content.
-   * @param  {string=} query.type - Filter by this type (Entry or Asset)
-   * @param  {string=} query.content_type - Filter by this content type id
-   * @param  {boolean=} query.resolveLinks - When true, links to other Entries or Assets are resolved. Default: true.
-   * @return {Promise<Sync.SyncCollection>} Promise for the collection resulting of a sync operation
-   * @example
-   * client.sync()
-   * .then((response) => console.log(response.entries, response.assets, response.nextSyncToken))
-   */
-  return function sync (query = {}) {
-    const resolveLinks = shouldLinksResolve(query)
-    return pagedSync(http, query, resolveLinks)
-  }
-}
-
 /**
  * This module retrieves all the available pages for a sync operation
  * @private
@@ -65,7 +43,7 @@ export default function createSyncRequester (http, shouldLinksResolve) {
  * @param {boolean} resolveLinks - If links should be resolved
  * @return {Promise<SyncCollection>}
  */
-function pagedSync (http, query, resolveLinks) {
+export default function pagedSync (http, query, resolveLinks) {
   if (!query || (!query.initial && !query.nextSyncToken)) {
     throw new Error('Please provide one of `initial` or `nextSyncToken` parameters for syncing')
   }
