+++ test/runner-browser.js
@@ -2,7 +2,7 @@
 // https://github.com/tmcw-up-for-adoption/karma-tap/issues/10
 require('./unit/contentful-test.js')
 require('./unit/create-contentful-api-test.js')
-require('./unit/create-sync-requester-test.js')
+require('./unit/paged-sync-test.js')
 require('./unit/entities/asset-test.js')
 require('./unit/entities/content-type-test.js')
 require('./unit/entities/entry-test.js')
